requirements.txt
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0
tensorboard>=2.12.0